requirements.txt
joblib==1.5.1
lightgbm==4.6.0
matplotlib==3.8.3
numpy==2.3.1
opencv_python==4.11.0.86
optuna==4.3.0
pandas==2.3.0
Pillow==11.2.1
scikit_learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
skimage==0.0
